refactor: ディレクトリ構成を変更した
