--- src/app/page.tsx
@@ -1,7 +1,0 @@
-import { Index } from '@/components/Home';
-
-const Welcome = () => {
-  return <Index />;
-};
-
-export default Welcome;
