+++ src/components/ui/Noise/index.tsx
